// front/src/MODELS/activityGenerationPrompt.ts
export interface IActivityGenerationPrompt {
    _id: string
    overview: string
    affordancesConstraints: string
    title: string
    constraint: string
    intent: string
    extensions: string
    scaffolding: string
    playerGroupSizes: string
    equipmentNeeded: string
    rules: string
    scoringSystem: string
    winCondition: string
    finalGuidelines: string
    createdAt: Date
    updatedAt: Date
}